Added Progressions Screen and Edit/Add Progression
Chords are only displayed at the moment and not editable (should be part
of a grid in the future).
Progressions Screen shows all saved progressions and allows the user to
edit or remove existing progressions and also add new ones to the list.
Edit and Add both both navigate to the same page which incudes a form,
and its fields are filled with the existing data in case of editing.